requirements.txt
matplotlib~=3.4.2
pandas~=1.3.0
numpy~=1.21.0
scipy~=1.7.0
jupyter
lmfit~=1.0.3
Pillow~=8.3.1